+++ projeto19-drivenpass/src/repositories/safeNoteRepository.ts
@@ -12,6 +12,7 @@ async function findById(id: number, userId: number) {
         where: { id }
     });
     if (!note) throw { type: "not_found" };
+    if (note.userId !== userId) throw { type: "this information doesn't belong to you" };
 
     return note;
 };
@@ -23,6 +24,10 @@ async function findAll(userId: number) {
 };
 
 async function deleteNote(id: number, userId: number) {
+    const note = await prisma.safeNote.findUnique({
+        where: { id },
+    });
+    if (note.userId !== userId) throw { type: "this information doesn't belong to you" };
     await prisma.safeNote.delete({
         where: { id }
     });
